requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blessings==1.7
bokeh==2.4.2
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.4
ConfigSpace==0.5.0
cycler==0.10.0
Cython==0.29.28
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
dill==0.3.4
distinctipy==1.2.1
einops==0.4.1
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.4.0
future==0.18.2
gast==0.3.3
google-api-core==1.31.1
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.32.0
gym==0.17.2
h5py==2.10.0
higher==0.2.1
hiredis==2.0.0
hpbandster==0.7.4
hyperopt==0.2.5
idna==2.10
importlib-resources==5.7.1
ipykernel==5.5.3
ipython @ file:///tmp/build/80754af9/ipython_1614615245721/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2==3.1.1
joblib==0.16.0
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213311222/work
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazydocs==0.4.8
learn2learn==0.1.3
llvmlite==0.38.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.2
multidict==5.1.0
netifaces==0.11.0
networkx==2.5.1
nltk==3.5
nmslib==2.1.1
numba==0.55.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.13
opencensus-context==0.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.1
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
patsy==0.5.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py-spy==0.3.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pydantic==1.8.2
pyglet==1.5.0
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pynndescent==0.5.6
pyparsing==2.4.7
Pyro4==4.82
pyrsistent==0.18.0
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytorch-lightning==1.2.10
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
ray==1.5.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.5.2
seaborn==0.11.1
serpent==1.40
six==1.15.0
statsmodels==0.13.2
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-datasets==4.5.2
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-metadata==1.7.0
tensorflow-similarity==0.15.8
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.6.0
torchmetrics==0.2.0
torchsummary==1.5.1
torchvision==0.7.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.48.2
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typer==0.4.1
typing-extensions==4.2.0
umap-learn==0.5.3
urllib3==1.25.10
vector-quantize-pytorch==0.6.0
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.4.1
yarl==1.6.3
zipp==3.8.0
